Update format of all footprints to KiCad8
Even though KiCad8 can read older symbol files, this change is needed
to make future additions to the library easy to review.
KiCad8 introduced a new way of indenting. Even when changing just a
small text, that would result in a huge diff when saving the KiCad7 file
with KiCad8. Those diffs are hard to read making the life of the library
team difficult.