Delete LED_Rectangular_W3.9mm_H1.8mm_FlatTop
The footprint seems to be the same as for
LED_Rectangular_W3.9mm_H1.8mm, and that one is already "flat enough".

The datasheet points to L-2774GD which officially has a round top.
Maybe perfectly rectangular ones with a flat top are being sold,
but it feels strange to have a duplicate footprint without even a 3D model
and relatively unusual name.