Rename diodes D_*_AK and D_*KA to LargeAnode & LargeCathode
The previous naming e.g. `D_MicroSMP_AK` was wrong. For all of the 6 changed
footprints, Pin1 was cathode, as it should be according to KLC. But the name
`AK` suggests that pin1=A, pin2=K which is not correct.
Instead the difference between those footprints is which of the asymmetrical
pads is the bigger one. Thus the new names have the `LargeAnode` and
`LargeCathode` suffix. Cathode is always Pin1. The description of those
footprints has been updated as well.

Also 'Kathode' was changed into 'Cathode' for the affected footprints. Because
its the proper english naming. The Pin names however do stay 'K'.

Closes: #2706
See https://gitlab.com/kicad/libraries/kicad-footprints/-/merge_requests/4450